% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_to_ph.R
\name{h_to_ph}
\alias{h_to_ph}
\title{Convert H to pH}
\usage{
h_to_ph(h)
}
\arguments{
\item{h}{Hydrogen ion concentration (nmol/kg)}
}
\value{
pH value
}
\description{
Convert H to pH
}
\examples{
h_to_ph(16)
}
