% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_to_h.R
\name{ph_to_h}
\alias{ph_to_h}
\title{Convert pH to H}
\usage{
ph_to_h(ph)
}
\arguments{
\item{ph}{pH value}
}
\value{
Hydrogen ion concentration (nmol/kg)
}
\description{
Convert pH to H
}
\examples{
ph_to_h(7.8)
}
